// src/frontend/src/backend.d.ts
export enum Rarity {
    Rare = "Rare",
    Uncommon = "Uncommon",
    Legendary = "Legendary",
    Common = "Common"
}